Fix function application for lambda expressions
Fixes (nested) application for lambdas.

There's now a lot of duplication in both the parsers (unavoidable?) and the
eval_tree function, and the :__apply case of eval_tree. Need to work to reduce
that.

Spent a lot of time troubleshooting the failure to look up variables in the
environment. In the end, there was a recursive call to which env wasn't being
passed.