Refactor some of the initial module/require code
This was going to introduce `provide`, but after getting one passing test,
they've all failed. Realized that there isn't a good mechanism in place with
this approach to 'compile' functions that are exported from a module that may
depend on a function that is private.

Opting for now to make all functions public when importing. There will be
namespace pollution; not sure best approach right now.