Build up right arrows in function inference; change arrow str
Building up arrows with a right fold is more correct, as terp functions are
curried and multi-argument functions apply the first argument and return a
function that takes the next argument until the function is saturated.

The string representation is changed to be lisp-like; the arrow is in the prefix
position now.